# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Medical billing",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/api/v1/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login, make auth",
                "description": "Make Auth in app",
                "operationId": "5db09a35f965c6ca7529ed3c3cae8e21",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "password": "helloworld",
                                    "email": "admin@billing.com"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Model User with roles and permission"
                    },
                    "401": {
                        "description": "User Unauthorized"
                    }
                }
            }
        },
        "/api/v1/auth/me": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "me, get info about user authenticated",
                "description": "Get information of user authenticate",
                "operationId": "1e5f382ba4fa0960d233875cab75a361",
                "responses": {
                    "200": {
                        "description": "Model User with roles and permission"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/auth/logout": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "logout, revoke auth",
                "description": "Make logout",
                "operationId": "8af307ba8ed9ad49ab1f1a5ad62bcd20",
                "responses": {
                    "200": {
                        "description": "Successfully logged out"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/auth/refresh_token": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "refresh_token, refresh token",
                "description": "Refresh Token",
                "operationId": "4bd176ce41a64986d5212d626fe4e2c6",
                "responses": {
                    "200": {
                        "description": "response a bearer token"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/user/": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create Users, endpoint to create users",
                "description": "Create User",
                "operationId": "f43914184ab28485d214312da63ba8c2",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "DBO": {
                                        "type": "date"
                                    },
                                    "sex": {
                                        "type": "char"
                                    },
                                    "firstName": {
                                        "type": "string"
                                    },
                                    "lastName": {
                                        "type": "char"
                                    },
                                    "middleName": {
                                        "type": "char"
                                    },
                                    "roles": {
                                        "type": "array",
                                        "items": {
                                            "property": null,
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "JessicaSmith",
                                    "password": "12345678",
                                    "email": "hola@mundo.com",
                                    "DBO": "2021-12-31",
                                    "sex": "F",
                                    "firstName": "Jhon",
                                    "lastName": "Due",
                                    "middleName": "Due",
                                    "roles": [
                                        "role1",
                                        "role2"
                                    ]
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Model User with roles and permission"
                    },
                    "500": {
                        "description": "Exception"
                    }
                }
            }
        },
        "/api/v1/user/send-email-rescue-pass": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Rescue password,Solicitude email to rescue password",
                "description": "Solicitude email to rescue password",
                "operationId": "889de411ac88da4bda44358e1cebbdbf",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "hola@mundo.com"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Rescue Email was sent"
                    },
                    "403": {
                        "description": "Error, a error have occurred"
                    },
                    "500": {
                        "description": "Exception"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/user/change-password/{token}": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Change Password, endpoint to change password to users",
                "description": "Change password users",
                "operationId": "dddda5f47c4abde2f6d1fdce59c25791",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "DBO": {
                                        "type": "date"
                                    },
                                    "sex": {
                                        "type": "char"
                                    },
                                    "firstName": {
                                        "type": "string"
                                    },
                                    "lastName": {
                                        "type": "char"
                                    },
                                    "middleName": {
                                        "type": "char"
                                    },
                                    "roles": {
                                        "type": "array",
                                        "items": {
                                            "property": null,
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "JessicaSmith",
                                    "password": "12345678",
                                    "email": "hola@mundo.com",
                                    "DBO": "2021-12-31",
                                    "sex": "F",
                                    "firstName": "Jhon",
                                    "lastName": "Due",
                                    "middleName": "Due",
                                    "roles": [
                                        "role1",
                                        "role2"
                                    ]
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Model User with roles and permission"
                    },
                    "500": {
                        "description": "Exception"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Authentication",
            "description": "API Endpoints of Authentication"
        },
        {
            "name": "Users",
            "description": "API Endpoints of Users"
        }
    ],
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}